Update controls preview to properly render themed assets.
The style preview card does not use assets which are overridden
by most themes. This results in an inaccurate preview of the
included controls, ie checkbox and radio buttons. Additionally,
the font size is cut off on the button in the preview. Increase
button height and decrease font size to compensate.

Ticket: CYNGNOS-874
Change-Id: I286f801afa7b6366ecf50395075240f33bc3d0ac